:recycle: Refactored all helper functinos
