+++ src/components/EpisodeDetails/helpers/index.js
@@ -1,0 +1,4 @@
+const removeAnyHTMLTagsFromString = (incomingStr) =>
+  incomingStr.replace(/(<([^>]+)>)/gi, '');
+
+export default removeAnyHTMLTagsFromString;
