+++ src/pages/MapPage.jsx
@@ -7,9 +7,7 @@ import EpisodeDetails from '../components/EpisodeDetails/EpisodeDetails';
 
 import { PodcastEpisodesContext } from '../contexts/PodcastEpisodesContext/PodcastEpisodesContext';
 
-import helperFunctions from '../helpers';
-
-const { isEmptyObject } = helperFunctions;
+import isEmptyObject from './helpers';
 
 const MapPage = () => {
   const { episodeDetails, audioPlayerVisibility } = useContext(
