+++ src/pages/helpers/index.js
@@ -1,0 +1,3 @@
+const isEmptyObject = (incomingObj) => Object.keys(incomingObj).length === 0;
+
+export default isEmptyObject;
