+++ src/utils/mediaQueries/index.js
@@ -1,0 +1,18 @@
+import { css } from 'styled-components';
+import breakpointsData from '../../data/breakpoints';
+
+const respondToDeviceWidth = Object.keys(breakpointsData).reduce(
+  (acc, label) => {
+    acc[label] = (...args) => css`
+      @media (min-width: ${breakpointsData[label]}) {
+        ${css(...args)};
+      }
+    `;
+    return acc;
+  },
+  {}
+);
+
+export default {
+  respondToDeviceWidth,
+};
